Solved CR from reviews. MAJOR CHANGE: Date to LocalDateTime
