+++ api/src/main/java/org/openmrs/module/sespct/api/PedidoService.java
@@ -3,6 +3,7 @@ package org.openmrs.module.sespct.api;
 import org.openmrs.api.OpenmrsService;
 import org.openmrs.module.sespct.api.model.Pedido;
 
+import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
 
@@ -18,7 +19,7 @@ public interface PedidoService extends OpenmrsService {
 	
 	List<Pedido> getPedidosByEstado(String estado);
 	
-	List<Pedido> getPedidosByDateRange(Date startDate, Date endDate);
+	List<Pedido> getPedidosByDateTimeRange(LocalDateTime startDateTime, LocalDateTime endDateTime);
 	
 	void deletePedido(Pedido pedido);
