+++ api/src/main/java/org/openmrs/module/sespct/api/model/DadosLaboratorioCargaViral.java
@@ -1,8 +1,10 @@
 package org.openmrs.module.sespct.api.model;
 
 import org.openmrs.BaseOpenmrsData;
+import org.openmrs.module.sespct.api.util.LocalDateTimeAttributeConverter;
 
 import javax.persistence.*;
+import java.time.LocalDateTime;
 import java.util.Date;
 
 @Entity
@@ -19,8 +21,8 @@ public class DadosLaboratorioCargaViral extends BaseOpenmrsData {
 	private Pedido pedido;
 	
 	@Column(name = "data_exame")
-	@Temporal(TemporalType.DATE)
-	private Date data;
+	@Convert(converter = LocalDateTimeAttributeConverter.class)
+	private LocalDateTime data;
 	
 	@Column(name = "carga_viral")
 	private Long cargaViral;
@@ -43,11 +45,11 @@ public class DadosLaboratorioCargaViral extends BaseOpenmrsData {
 		this.pedido = pedido;
 	}
 	
-	public Date getData() {
+	public LocalDateTime getData() {
 		return data;
 	}
 	
-	public void setData(Date data) {
+	public void setData(LocalDateTime data) {
 		this.data = data;
 	}
