+++ api/src/main/resources/messages.properties
@@ -30,4 +30,5 @@ sespct.search.status.NAO_PROCESSADO=Not Processed
 sespct.search.ncft=NCFT:
 sespct.search.nid=NID:
 sespct.search.submit=Search
+sespct.error.loadingData =Error Loading Data
