+++ api/src/main/resources/messages_pt.properties
@@ -30,3 +30,4 @@ sespct.search.status.NAO_PROCESSADO=Não processado
 sespct.search.ncft=NCFT:
 sespct.search.nid=NID:
 sespct.search.submit=Pesquisar
+sespct.error.loadingData =Erro a carregar dados
